Move json files to config folder
